requirements.dev.txt
flake8==7.3.0
psycopg2-binary==2.9.11